Added Main Class and decrypter works now.
The main class was added to run encrypter and decrypter from,

I'm not sure if it's necessary so I might delete it later.

I also completed decrypter, I had to move the digit swap to after the

reversal of the modulus 10 (which) shifted the digits forward 3.

All the files are now connected to the Main Class.